requirements.txt
fastapi==0.104.1
uvicorn==0.30.6
requests==2.32.3
feedparser==6.0.11
python-dateutil==2.9.0.post0
pydantic==1.10.15
python-dotenv==1.0.1
rich==13.9.4
